[refactor] Created lib dir for utilities;Refactored API and S3 code
