+++ drizzle.config.ts
@@ -5,7 +5,7 @@ config({ path: '.env.local' });
 
 export default defineConfig({
   out: './src/drizzle',
-  schema: './src/app/db/schema.ts',
+  schema: './src/app/lib/db/schema.ts',
   dialect: 'postgresql',
   dbCredentials: {
     url: process.env.DATABASE_URL!,
